fix: Avoid the risk of long transactions and move db operations to the appropriate service layer
