+++ api/controllers/console/apikey.py
@@ -6,7 +6,7 @@ from sqlalchemy.orm import Session
 from werkzeug.exceptions import Forbidden
 
 from extensions.ext_database import db
-from libs.api_token_cache import ApiTokenCache
+from services.api_token_service import ApiTokenCache
 from libs.helper import TimestampField
 from libs.login import current_account_with_tenant, login_required
 from models.dataset import Dataset
