+++ api/tasks/remove_app_and_related_data_task.py
@@ -14,7 +14,7 @@ from sqlalchemy.orm import sessionmaker
 from configs import dify_config
 from core.db.session_factory import session_factory
 from extensions.ext_database import db
-from libs.api_token_cache import ApiTokenCache
+from services.api_token_service import ApiTokenCache
 from libs.archive_storage import ArchiveStorageNotConfiguredError, get_archive_storage
 from models import (
     ApiToken,
